modularized all external plugins
