+++ configure.ac
@@ -553,7 +553,6 @@ AC_CONFIG_FILES([
     netdata.spec
     python.d/Makefile
     node.d/Makefile
-    plugins.d/Makefile
     src/api/Makefile
     src/backends/graphite/Makefile
     src/backends/json/Makefile
@@ -564,19 +563,23 @@ AC_CONFIG_FILES([
     src/health/Makefile
     src/libnetdata/Makefile
     src/Makefile
-    src/plugins/apps.plugin/Makefile
     src/plugins/checks.plugin/Makefile
     src/plugins/freebsd.plugin/Makefile
     src/plugins/idlejitter.plugin/Makefile
     src/plugins/linux-cgroups.plugin/Makefile
     src/plugins/linux-diskspace.plugin/Makefile
-    src/plugins/linux-freeipmi.plugin/Makefile
     src/plugins/linux-nfacct.plugin/Makefile
     src/plugins/linux-proc.plugin/Makefile
     src/plugins/linux-tc.plugin/Makefile
     src/plugins/macos.plugin/Makefile
     src/plugins/Makefile
     src/plugins/plugins.d.plugin/Makefile
+    src/plugins/plugins.d.plugin/apps.plugin/Makefile
+    src/plugins/plugins.d.plugin/charts.d.plugin/Makefile
+    src/plugins/plugins.d.plugin/fping.plugin/Makefile
+    src/plugins/plugins.d.plugin/freeipmi.plugin/Makefile
+    src/plugins/plugins.d.plugin/node.d.plugin/Makefile
+    src/plugins/plugins.d.plugin/python.d.plugin/Makefile
     src/plugins/statsd.plugin/Makefile
     src/registry/Makefile
     src/streaming/Makefile
