+++ src/plugins/plugins.d.plugin/Makefile.am
@@ -3,3 +3,15 @@
 AUTOMAKE_OPTIONS = subdir-objects
 MAINTAINERCLEANFILES = Makefile.in
 
+SUBDIRS = \
+    apps.plugin \
+    charts.d.plugin \
+    freeipmi.plugin \
+    fping.plugin \
+    node.d.plugin \
+    python.d.plugin \
+    $(NULL)
+
+dist_noinst_DATA = \
+    README.md \
+    $(NULL)
