+++ src/plugins/plugins.d.plugin/apps.plugin/Makefile.am
