+++ src/plugins/plugins.d.plugin/apps.plugin/apps_plugin.c
@@ -6,7 +6,7 @@
  * Released under GPL v3+
  */
 
-#include "../../libnetdata/libnetdata.h"
+#include "../../../libnetdata/libnetdata.h"
 
 // ----------------------------------------------------------------------------
