+++ src/plugins/plugins.d.plugin/charts.d.plugin/charts.d.plugin.in
