+++ src/plugins/plugins.d.plugin/fping.plugin/Makefile.am
@@ -4,14 +4,13 @@ AUTOMAKE_OPTIONS = subdir-objects
 MAINTAINERCLEANFILES = Makefile.in
 
 CLEANFILES = \
-	alarm-notify.sh \
+	fping.plugin \
 	$(NULL)
 
-alarm-notify.sh: alarm-notify.sh.in
+fping.plugin: fping.plugin.in
 	if sed \
 		-e 's#[@]configdir_POST@#$(configdir)#g' \
 		-e 's#[@]libconfigdir_POST@#$(libconfigdir)#g' \
-		-e 's#[@]cachedir_POST@#$(cachedir)#g' \
 		$< > $@.tmp; then \
 		mv "$@.tmp" "$@"; \
 	else \
@@ -20,11 +19,9 @@ alarm-notify.sh: alarm-notify.sh.in
 	fi
 
 dist_plugins_SCRIPTS = \
-	alarm-notify.sh \
-	alarm-email.sh \
-	alarm-test.sh \
+	fping.plugin \
 	$(NULL)
 
 dist_noinst_DATA = \
-	alarm-notify.sh.in \
+	fping.plugin.in \
 	$(NULL)
