+++ src/plugins/plugins.d.plugin/fping.plugin/fping.plugin.in
