+++ src/plugins/plugins.d.plugin/freeipmi.plugin/Makefile.am
