+++ src/plugins/plugins.d.plugin/node.d.plugin/node.d.plugin.in
@@ -40,7 +40,7 @@ var util = require('util');
 var http = require('http');
 var path = require('path');
 var extend = require('extend');
-var netdata = require('netdata');
+var netdata = require('../../../netdata');
 
 
 // --------------------------------------------------------------------------------------------------------------------
